requirements.txt
django>=4.2
mongoengine>=0.28
django-cors-headers>=4.0
Pillow>=10.0
pix2tex>=0.1.0
boto3>=1.28
PyYAML>=6.0
python-docx>=1.1
latex2mathml>=3.77
qdrant-client
requests>=2.28.0
volcengine-python-sdk[ark]
langchain-openai>=0.2.0
langchain-core>=0.3.0
